+++ server/repositories/WalletRepository.spec.js
@@ -75,6 +75,8 @@ describe('WalletRepository', () => {
       expect(query.sql).match(
         /select.*wallet.*where.*actor_wallet_id.*request_type.*/,
       );
+      expect(query.sql.match(/display_name/g)).to.have.length(3);
+      expect(query.sql.match(/cover_url/g)).to.have.length(3);
       query.response([{ id: 1 }]);
     });
     const entity = await walletRepository.getAllWallets(uuid.v4());